+++ CHANGELOG.md
@@ -1,14 +1,14 @@
 # Unreleased
 
 - **Breaking:** Added `make_current_surfaceless(self)` for `{Possibly,Not}CurrentGlContext`.
+- **Breaking:** CGL (macOS) no longer requests an `NSOpenGLProfile` unless `Api::OPENGL` was
+  requested via `ConfigTemplateBuilder::with_api`.
 - Updated `objc2` frameworks dependency to `v0.3.2`.
 - Use `objc2-open-gl` instead of `cgl` dependency.
 - Bump MSRV from `1.71` to `1.85`.
 - Fixed EGL's robustness detection without extension, but on EGL 1.5.
 - Fixed building docs on docs.rs.
 - Fixed WGL's unreliable TRANSPARENT_ARB false negatives by returning None
-- Added `ConfigTemplateBuilder::skip_cgl_profile` to allow skipping CGL's `NSOpenGLProfile`
-  (allowing an older openGL version with extensions to work on macOS).
 
 # Version 0.32.3